pyproject.toml
[project]
name = "agi3"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "pynacl",
]
